% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-ed2-xml.R
\name{as_pft_list}
\alias{as_pft_list}
\title{Convert `data.frame` to PFT list suitable for `write_ed2_xml` input}
\usage{
as_pft_list(dat)
}
\arguments{
\item{dat}{`data.frame` with PFT trait data}
}
\value{
Named nested list of PFTs and trait values
}
\description{
Convert `data.frame` to PFT list suitable for `write_ed2_xml` input
}
\author{
Alexey Shiklomanov
}
