% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodel-rtm.R
\name{beta0}
\alias{beta0}
\title{Beam backscatter}
\usage{
beta0(theta, orient)
}
\arguments{
\item{theta}{Projection angle (degrees)}

\item{orient}{Leaf orientation factor (-1, 1)}
}
\description{
Beam backscatter
}
