% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bety.R
\name{bety}
\alias{bety}
\title{Shortcut to open a BETY connection}
\usage{
bety(user = "bety", password = "bety", host = "localhost",
  port = 7990)
}
\arguments{
\item{user, password, host, port}{Configuration options for BETY
connection. Defaults are configured for A. Shiklomanov's PNNL
instance.}
}
\value{
DBI database connection object.
}
\description{
Shortcut to open a BETY connection
}
