% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bety.R
\name{bety_tables}
\alias{bety_tables}
\alias{pfts_priors}
\alias{pfts_species}
\title{Shortcut access functions for BETY table joins}
\usage{
pfts_priors(pft_names = pfts("bety_name"), con = bety(), collect = TRUE)

pfts_species(pft_names = pfts("bety_name"), con = bety(), collect = TRUE)
}
\arguments{
\item{pft_names}{Character vector of PFT names. Defaults to
`bety_name` column of [pfts()]. If `NULL`, don't filter by PFT name.}

\item{con}{Database connection. Defaults to [bety()].}

\item{collect}{Whether or not to call [dplyr::collect()] on the
output. Default = `TRUE`.}
}
\description{
Shortcut access functions for BETY table joins
}
\details{
- `pfts_priors` -- Prior distributions by PFT
- `pfts_species` -- PFT-species mapping
}
