% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-run-groups.R
\name{create_run_groups}
\alias{create_run_groups}
\title{Sort runs into clusters based on most dissimilar LAI}
\usage{
create_run_groups(dat, ..., n = 4)
}
\arguments{
\item{dat}{Input `data.frame`}

\item{...}{Ignored (necessary for this to work with
[dplyr::group_modify()])}
}
\value{
`data.frame` containing `run_id` and `cluster`
}
\description{
Sort runs into clusters based on most dissimilar LAI
}
\author{
Alexey Shiklomanov
}
