% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_connection.R
\name{default_connection}
\alias{default_connection}
\title{Default PEcAn database connection}
\usage{
default_connection()
}
\value{
DBI connection object
}
\description{
Default PEcAn database connection
}
\author{
Alexey Shiklomanov
}
