% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_history_file.R
\name{download_history_file}
\alias{download_history_file}
\title{Download a history output file for a given workflow}
\usage{
download_history_file(workflow_id, year, month, day, hour,
  outdir = file.path("analysis", "data", "model_output", "histfiles"))
}
\arguments{
\item{workflow_id}{ID of PEcAn workflow (from BETY `workflows` table)}

\item{year}{Output year}

\item{month}{Output month}

\item{day}{Output day}

\item{hour}{Output hour}

\item{outdir}{Target directory for storing files. Default is
`analysis/data/model_output/histfiles`}
}
\value{
Full path to downloaded file
}
\description{
Download a history output file for a given workflow
}
\author{
Alexey Shiklomanov
}
