% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecan_meta_analysis.R
\name{draw_traits}
\alias{draw_traits}
\title{Draw traits from data frame describing prior or posterior}
\usage{
draw_traits(data, ndraws = 5000)
}
\arguments{
\item{data}{`data.frame` describing distributions, with columns
`distn` (name of distribution function), `parama` and `paramb`
(distribution parameters)}

\item{ndraws}{Number of random draws for calculating summary stats
(default = 5000).}
}
\value{
Tidy `data.frame` containing posterior information, draws,
  and summary statistics
}
\description{
Draw traits from data frame describing prior or posterior
}
