% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed-default-params.R
\name{ed_default_params}
\alias{ed_default_params}
\title{ED2 default parameters}
\usage{
ed_default_params()
}
\value{
Tidy `data.frame` of ED2 default parameter values
}
\description{
ED2 default parameters
}
\author{
Alexey Shiklomanov
}
