% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow_workflow.R
\name{follow_workflow}
\alias{follow_workflow}
\title{Continuously stream the `workflow.Rout` file of a workflow until it
is finished}
\usage{
follow_workflow(workflow_id, refresh = 2,
  error_rxp = "> proc.time\\\\(\\\\)", start_at = 1, ...)
}
\arguments{
\item{workflow_id}{Workflow ID of workflow to monitor}

\item{...}{Additional arguments to `pecanapi::output_url`}
}
\value{
Workflow output as character vector, invisibly
}
\description{
Continuously stream the `workflow.Rout` file of a workflow until it
is finished
}
\author{
Alexey Shiklomanov
}
