% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieved-paths.R
\name{fst_cohort}
\alias{fst_cohort}
\alias{fst_pft}
\alias{fst_scalar}
\alias{fst_soil}
\alias{fst_cohort_monthly}
\alias{fst_pft_monthly}
\alias{fst_scalar_monthly}
\alias{fst_soil_monthly}
\title{Shortcuts for processed output files}
\usage{
fst_cohort(root = here::here())

fst_pft(root = here::here())

fst_scalar(root = here::here())

fst_soil(root = here::here())

fst_cohort_monthly(root = here::here())

fst_pft_monthly(root = here::here())

fst_scalar_monthly(root = here::here())

fst_soil_monthly(root = here::here())
}
\arguments{
\item{root}{Project root directory. Default = `here::here()`.}
}
\value{
`fst::fst` object pointing to dataset
}
\description{
Shortcuts for processed output files
}
\author{
Alexey Shiklomanov
}
