% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvector.R
\name{fvector}
\alias{fvector}
\alias{fmatrix}
\title{Create a variable-index vector or matrix}
\usage{
fvector(indices)

fmatrix(irows, icols)
}
\arguments{
\item{indices}{Vector indices}

\item{irows, icols}{Indices of rows or columns}
}
\value{
Numeric vector with "custom" indices
}
\description{
Create a variable-index vector or matrix
}
\examples{
x <- fvector(0:2)
x[] <- c(5, 7, 9)
x
x[0]
# This matrix is square
y <- fmatrix(0:5, 1:6)
}
\author{
Alexey Shiklomanov
}
