% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bety.R
\name{get_ids}
\alias{get_ids}
\title{Retrieve ID from a BETY table given a name}
\usage{
get_ids(values, table, value_col, id_col = "id", con = bety())
}
\arguments{
\item{values}{Vector of values for which to retrieve IDs}

\item{table}{Name of table where IDs are stored}

\item{value_col}{Name of the column corresponding to the values}

\item{id_col}{Name of the column containing the ID. Default =
`"id"`}

\item{con}{Database connection. Defaults to [bety()].}
}
\value{
Named numeric vector of IDs
}
\description{
Retrieve ID from a BETY table given a name
}
