% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_lai.R
\name{get_monthly_lai}
\alias{get_monthly_lai}
\title{Read monthly LAI an ED `-T-` file}
\usage{
get_monthly_lai(workflow_id, year, startmonth = 1, startday = 1)
}
\arguments{
\item{workflow_id}{ID of PEcAn workflow (from BETY `workflows` table)}

\item{year}{Target year}

\item{startmonth}{Start month for reading in data. Default = 1.}

\item{startday}{Start day for reading in data. Default = 1.}
}
\value{
`data.frame` of monthly LAI values by PFT
}
\description{
Read monthly LAI an ED `-T-` file
}
\author{
Alexey Shiklomanov
}
