% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bety.R
\name{get_pft_ids}
\alias{get_pft_ids}
\alias{get_species_ids}
\alias{get_variable_ids}
\title{Extract BETY PFT, species, or variable IDs given a vector of names}
\usage{
get_pft_ids(pft_names, con = bety())

get_species_ids(species_names, con = bety())

get_variable_ids(variable_names, con = bety())
}
\arguments{
\item{pft_names}{Character vector of PFT names. Defaults to
`bety_name` column of [pfts()].}

\item{con}{Database connection. Defaults to [bety()].}

\item{species_names}{Character vector of species names}

\item{variables}{Character vector of variable names}
}
\value{
Numeric vector of BETY IDs for given PFTs or species
}
\description{
Extract BETY PFT, species, or variable IDs given a vector of names
}
