% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodel-rtm.R
\name{gfunction}
\alias{gfunction}
\alias{kfunction}
\title{Relative (G) or absolute (K) projected leaf area in the target direction}
\usage{
gfunction(theta, orient, theta_rad = theta * pi/180)

kfunction(theta, orient, theta_rad = theta * pi/180)
}
\arguments{
\item{theta}{Projection angle (degrees)}

\item{orient}{Leaf orientation factor (-1, 1)}

\item{theta_rad}{Projection angle (radians)}
}
\description{
Relative (G) or absolute (K) projected leaf area in the target direction
}
