% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodel-rtm.R
\name{integrate_light}
\alias{integrate_light}
\title{Integrate light levels from flux estimates}
\usage{
integrate_light(down0, down, up)
}
\arguments{
\item{down0}{Cohort-level direct downwelling flux}

\item{down}{Cohort-level diffuse downwelling flux}

\item{up}{Cohort-level diffuse upwelling flux}
}
\value{
List containing overall light level, and levels of direct and diffuse light
}
\description{
Integrate light levels from flux estimates
}
\author{
Alexey Shiklomanov
}
