% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecan-ed2.R
\name{listToXml}
\alias{listToXml}
\title{Convert List to XML}
\usage{
listToXml(item, tag)
}
\arguments{
\item{item}{object to be converted.
Despite the function name, need not actually be a list}

\item{tag}{xml tag}
}
\value{
xmlNode
}
\description{
Can convert list or other object to an xml object using xmlNode
}
\author{
David LeBauer, Carl Davidson, Rob Kooper
}
