% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodel-rtm-ms.R
\name{ms_tau_diffuse}
\alias{ms_tau_diffuse}
\title{Transmissivity to diffuse radiation for the multiple-scatter model}
\usage{
ms_tau_diffuse(L, orient)
}
\arguments{
\item{L}{Leaf (or total) area index}

\item{orient}{Leaf orientation factor (-1, 1)}
}
\description{
Transmissivity to diffuse radiation for the multiple-scatter model
}
