% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodel-rtm-ms.R
\name{multi_scatter}
\alias{multi_scatter}
\title{Multiple-scatter model}
\usage{
multi_scatter(ipft, L, lr, lt, orient, C = 1, theta = 15,
  S0_beam = 0.8, S0_diff = 1 - S0_beam, alb_ground = 0.1)
}
\arguments{
\item{ipft}{Integer indices of PFTs (`icohort`)}

\item{L}{Leaf (or total) area index}

\item{lr}{Leaf reflectance (`ipft`)}

\item{lt}{Leaf transmittance (`ipft`)}

\item{orient}{Leaf orientation factor (-1, 1)}

\item{C}{Crown area index (default = 1)}

\item{theta}{Projection angle (degrees)}
}
\value{
List containin cohort-level downward beam radiation, and downward and
  upward diffuse radiation.
}
\description{
Cohorts are arranged from bottom (`i = 1`) to top (`i = ncohort`)
}
\author{
Alexey Shiklomanov
}
