% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_configs.R
\name{parse_configs}
\alias{parse_configs}
\title{Parse ED `config.xml` files}
\usage{
parse_configs(workflow_id, con = default_connection(), pfts = NULL)
}
\arguments{
\item{workflow_id}{Target workflow ID}

\item{con}{Database connection. Default is [default_connection()].}

\item{pfts}{PFTs to use. If `NULL`, use default PFTs: Early, mid,
and late hardwoods, and pine, in that order.}
}
\value{
`data.frame` of trait values for that run
}
\description{
Parse ED `config.xml` files
}
\author{
Alexey Shiklomanov
}
