% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_notes.R
\name{parse_notes}
\alias{parse_notes}
\title{Convert a regular `notes` entry to a `data.frame`}
\usage{
parse_notes(notes)
}
\arguments{
\item{notes}{Notes entry such as that created by
[[run_ed_ensemble()]], as a single character}
}
\value{
`data.frame` with columns for each setting
}
\description{
Convert a regular `notes` entry to a `data.frame`
}
\author{
Alexey Shiklomanov
}
