% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pb_along}
\alias{pb_along}
\title{Create progress bar object with length equal to target object}
\usage{
pb_along(x, ...)
}
\arguments{
\item{x}{Object "along" which to create progress bar}

\item{...}{Additional arguments to `progress::progress_bar$new()`}
}
\value{
`progress` object (see [progress::progress_bar()]).
}
\description{
Create progress bar object with length equal to target object
}
\author{
Alexey Shiklomanov
}
