% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pbn}
\alias{pbn}
\title{Create a progress bar with total = `n`}
\usage{
pbn(n, ...)
}
\arguments{
\item{n}{Number of progress ticks}
}
\value{
Progress bar object
}
\description{
Create a progress bar with total = `n`
}
