% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecan_meta_analysis.R
\name{pecan_ma_pft}
\alias{pecan_ma_pft}
\title{Run PEcAn trait meta analysis for a specific PFT}
\usage{
pecan_ma_pft(con, pft)
}
\arguments{
\item{con}{DBI connection object to PEcAn database}

\item{pft}{Plant functional type name}
}
\value{
Raw meta analysis output (see [PEcAn.MA::pecan.ma()])
}
\description{
Run PEcAn trait meta analysis for a specific PFT
}
\author{
Alexey Shiklomanov
}
