% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pft.R
\name{pfts}
\alias{pfts}
\title{PFT information}
\usage{
pfts(col = NULL)
}
\arguments{
\item{col}{Column to return. If `NULL` (default), return `data.frame`}
}
\value{
If `is.null(col)`, `data.frame` containing PEcAn PFT name
  (`bety_name`), formatted name (`pft`), short name (`shortname`),
  ED PFT number (`num`), and color (`color`).
}
\description{
PFT information
}
\author{
Alexey Shiklomanov
}
