% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodel-rtm.R
\name{phi1_f}
\alias{phi1_f}
\alias{phi2_f}
\title{Leaf orientation coefficients}
\usage{
phi1_f(orient)

phi2_f(orient, phi1 = phi1_f(orient))
}
\arguments{
\item{orient}{Leaf orientation factor (-1, 1)}
}
\description{
Leaf orientation coefficients
}
