% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qdate}
\alias{qdate}
\title{Create datetime object from isolated pieces}
\usage{
qdate(year, month = 1, day = 1, hour = 0, minute = 0, second = 0, tz = "UTC")
}
\arguments{
\item{year}{Year (numeric)}

\item{month}{Month (numeric, default = 1)}

\item{day}{Day of the month (numeric, default = 1)}

\item{hour}{Hour (numeric, default = 0)}

\item{minute}{Minute (numeric, default = 0)}

\item{second}{Second (numeric, default = 0)}

\item{tz}{Time zone (character, default = "UTC")}
}
\value{
An object of class \code{"\link[base]{POSIXct}"}.
}
\description{
Thin wrapper around [base::ISOdatetime()] that provides default
arguments. All arguments are vectorized.
}
\examples{
# Create dates given only the years (midnight, January 1)
qdate(2005:2010)
# Every September 5 (midnight)
qdate(2005:2010, 9, 5)
# ...at noon
qdate(2005:2010, 9, 5, 12)
}
\author{
Alexey Shiklomanov
}
