% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cohort_history.R
\name{read_cohort_history}
\alias{read_cohort_history}
\title{Read cohort variables from history file}
\usage{
read_cohort_history(workflow_id, year, month, day = 1, hour = 0,
  vars = NULL, .default_vars = c("PFT", "HITE", "LAI_CO", "VM0",
  "SLA"))
}
\arguments{
\item{workflow_id}{ID of PEcAn workflow (from BETY `workflows` table)}

\item{year}{Output year}

\item{month}{Output month}

\item{day}{Output day of month. Default = 1.}

\item{hour}{Output hour of day. Default = 0 (midnight).}

\item{...}{Additional}

\item{varlist}{List of variables to retrieve.}
}
\value{
`data.frame` of cohort output
}
\description{
Read cohort variables from history file
}
\author{
Alexey Shiklomanov
}
