% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_configxml.R
\name{read_configxml}
\alias{read_configxml}
\title{Read ED2 parameter file (`config.xml`) into a `data.frame`}
\usage{
read_configxml(path, filename = "config.xml")
}
\arguments{
\item{path}{Path to run directory containing XML config file}

\item{filename}{Name of config file. Default = `config.xml`}
}
\value{
Wide `data.frame` containing parameters for each PFT.
}
\description{
Read ED2 parameter file (`config.xml`) into a `data.frame`
}
\author{
Alexey Shiklomanov
}
