% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-ie-files.R
\name{read_e_file}
\alias{read_e_file}
\title{Read monthly (-E-) file}
\usage{
read_e_file(fname, .pb = NULL)
}
\arguments{
\item{fname}{File name}

\item{.pb}{Optional progress bar object}
}
\value{

}
\description{
Read monthly (-E-) file
}
\author{
Alexey Shiklomanov
}
