% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ed_output.R
\name{read_ed_t_file}
\alias{read_ed_t_file}
\title{Read ED output files}
\usage{
read_ed_t_file(filename)
}
\description{
Read the different kinds of ED output file (see Details).
}
\details{
ED can produce one of five kinds of output file:
- Daily averages (D)
- Monthly averages (M)
- Yearly averages (Y)
- Instantaneous fluxes, one file per year (T, for "tower")
- Monthly means of the diurnal cycle (Q) -- number of points is `1/FRQANL`
- History files (all variables needed for restart) (S) -- Restart
interval determined by `FRQHIS`
- "Fast" analysis files (F) -- Polygon-level averages at a custom
frequency (`FRQANL`)
- Observation output files (O) -- Same as F files, but only at
timesteps specified by `OBSTIME_DB`
}
