% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-ie-files.R
\name{read_efile_dir}
\alias{read_efile_dir}
\title{Read directory full of E files}
\usage{
read_efile_dir(outdir)
}
\arguments{
\item{outdir}{Output directory containing E files}
}
\value{
Nested `data.frame` containing the output directory and scalar,
  cohort, soil, and PFT results
}
\description{
Read directory full of E files
}
\author{
Alexey Shiklomanov
}
