% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-o-file.R
\name{read_i_cohort}
\alias{read_i_cohort}
\title{Read cohort data from `I` file}
\usage{
read_i_cohort(fname)
}
\arguments{
\item{fname}{File name}
}
\value{
`tibble` of cohort data
}
\description{
Read cohort data from `I` file
}
\author{
Alexey Shiklomanov
}
