% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_notes.R
\name{read_notes}
\alias{read_notes}
\title{Read the workflow notes column of a FoRTE run}
\usage{
read_notes(workflow_id, con = NULL)
}
\arguments{
\item{workflow_id}{Workflow ID of target workflow}

\item{con}{PEcAn database connection. If `NULL` (default), create a
connection using default settings.}
}
\value{
Named list describing the ED configuration of that workflow
}
\description{
Read the workflow notes column of a FoRTE run
}
\author{
Alexey Shiklomanov
}
