% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-ie-files.R
\name{read_o_file}
\alias{read_o_file}
\title{Read observed (-O-) file}
\usage{
read_o_file(fname, .pb = NULL)
}
\arguments{
\item{fname}{File name}

\item{.pb}{Optional progress bar object}
}
\value{

}
\description{
Read observed (-O-) file
}
\author{
Alexey Shiklomanov
}
