% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad_profile.R
\name{read_tidy_rad_profile}
\alias{read_tidy_rad_profile}
\title{Read the radiation profile from a history file and return as a tidy
`data.frame`}
\usage{
read_tidy_rad_profile(file)
}
\arguments{
\item{file}{Full path to history file}
}
\value{
`data.frame` with columns `workflow_id`, `datetime`,
  `cohort` (integer - tallest first), and one column per radiation flux.
}
\description{
Read the radiation profile from a history file and return as a tidy
`data.frame`
}
\author{
Alexey Shiklomanov
}
