% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-ed.R
\name{run_ed}
\alias{run_ed}
\title{Start an ED2 model run using `processx`}
\usage{
run_ed(
  casename,
  trait_values = list(),
  start_date = "1902-06-01",
  end_date = "1920-01-01",
  trait_plasticity = FALSE,
  multiple_scatter = FALSE,
  crown_model = FALSE,
  water_lim = TRUE,
  out_root = getOption("fortebaseline.ed_root"),
  ed_exe = fs::path(getOption("fortebaseline.ed_src_dir"), "ED", "build", "ed_2.1-opt"),
  ...
)
}
\arguments{
\item{casename}{Name}

\item{trait_values}{Named list of custom parameters}

\item{start_date, end_date}{Run start and end dates. Default is 1902-06-01 to
1920-01-01.}

\item{trait_plasticity}{Whether or not to enable the trait
plasticity scheme (default = `FALSE`)}

\item{multiple_scatter}{Whether or not to use the
multiple-scattering canopy RTM. If `FALSE` (default), use the
two-stream RTM.}

\item{crown_model}{Whether or not to use the finite canopy radius
model (default = `FALSE`)}

\item{water_lim}{Whether or not to run with water limitation to
photosynthesis (default = `TRUE`).}

\item{out_root}{Path to ED2 output root directory. Default is the option
`fortebaseline.ed_root`, and if unset, "ed2-output" in the project root
directory.}

\item{...}{Additional ED2IN namelist settings. Must be in all caps, and must
match settings in ED2IN.}
}
\value{
`processx` process for ED2 run.
}
\description{
Start an ED2 model run using `processx`
}
\author{
Alexey Shiklomanov
}
