% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-ed.R
\name{run_ed_maybe}
\alias{run_ed_maybe}
\title{Run ED2, but only if there isn't a run there already}
\usage{
run_ed_maybe(casename, overwrite = FALSE,
  out_root = getOption("fortebaseline.ed_root"), ...)
}
\arguments{
\item{casename}{Name}

\item{overwrite}{(Logical) If `TRUE`, delete the target output directory and
force a run.}

\item{out_root}{Path to ED2 output root directory. Default is the option
`fortebaseline.ed_root`, and if unset, "ed2-output" in the project root
directory.}

\item{...}{Additional ED2IN namelist settings. Must be in all caps, and must
match settings in ED2IN.}
}
\value{
List containing the output directory (`outdir`), the
  `processx::process` object (`p`), and a function for checking the current
  log status (`log`).
}
\description{
Run ED2, but only if there isn't a run there already
}
