% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfile.R
\name{runfiles}
\alias{runfiles}
\title{Retrieve all the `run` files from a particular workflow.}
\usage{
runfiles(workflow_id, target, con = default_connection())
}
\arguments{
\item{workflow_id}{ID of PEcAn workflow (from BETY `workflows` table)}

\item{con}{DBI database connection object for BETY database.
Default is output of [default_connection()].}
}
\value{
Character vector of URLs for reading or downloading files.
}
\description{
Retrieve all the `run` files from a particular workflow.
}
\author{
Alexey Shiklomanov
}
