% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{sensitivity_analysis}
\alias{sensitivity_analysis}
\title{One-at-a-time PEcAn sensitivity analysis}
\usage{
sensitivity_analysis(
  y,
  x,
  ymedian = median(y),
  xmedian = median(x),
  alpha = 0.05
)
}
\arguments{
\item{y}{Response variable}

\item{x}{Dependent variable}

\item{ymedian}{Median of response variable (default is calculated as
`median(y)`)}

\item{xmedian}{Median of dependent variable (default is calculated as
`median(x)`)}

\item{alpha}{Fraction of median around which to calculate
derivative for sensitivity. Default = 0.05}
}
\value{
`tibble` of sensitivity analysis results, with columns
`cv` (Coefficient of variation for `x`), `sensitivity`
(unnormalized sensitivity of `y` to `x` at median), `elasticity`
(normalized `sensitivity`), and `pvar` (prediction variance).
}
\description{
NOTE: Fitting all at once results in too few degrees of freedom.
Need to fit one at a time.
}
