% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pft.R
\name{set_pft}
\alias{set_pft}
\title{Set PFT from ED integer number}
\usage{
set_pft(i)
}
\arguments{
\item{i}{Vector of PFT integer numbers}
}
\value{
Factor of PFT names
}
\description{
Set PFT from ED integer number
}
\author{
Alexey Shiklomanov
}
