% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodel-rtm.R
\name{single_scatter_albedo}
\alias{single_scatter_albedo}
\title{Single-scattering albedo}
\usage{
single_scatter_albedo(theta, orient, theta_rad = theta * pi/180)
}
\arguments{
\item{theta}{Projection angle (degrees)}

\item{orient}{Leaf orientation factor (-1, 1)}

\item{theta_rad}{Projection angle (radians)}
}
\description{
Single-scattering albedo
}
