% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecan_meta_analysis.R
\name{summarize_ma}
\alias{summarize_ma}
\title{Summarize meta analysis output for multiple PFTs as a tidy `data.frame`}
\usage{
summarize_ma(result)
}
\arguments{
\item{result}{List of PFT-level meta-analysis results}
}
\value{
`data.frame` of summary statistics from meta analysis
}
\description{
Summarize meta analysis output for multiple PFTs as a tidy `data.frame`
}
