% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow_workflow.R
\name{tail_file}
\alias{tail_file}
\title{Follow the output of a file, similar to Linux's `tail -f`}
\usage{
tail_file(filename, refresh = 2, error_rxp = NULL, start_at = 1)
}
\arguments{
\item{filename}{File to follow}

\item{refresh}{Refresh interval in seconds. Default = 2.}

\item{error_rxp}{Regular expression to use for completion}

\item{start_at}{Line at which to start reading the file. Default = 1.}
}
\description{
Follow the output of a file, similar to Linux's `tail -f`
}
