% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecan_meta_analysis.R
\name{tidy_posterior}
\alias{tidy_posterior}
\title{Convert list of meta-analysis results to a tidy `data.frame` of the
posterior distribution.}
\usage{
tidy_posterior(ma_results, ndraws = 5000)
}
\arguments{
\item{ma_results}{List of meta-analysis results for multiple PFTs}

\item{ndraws}{Number of random draws for calculating summary stats
(default = 5000).}
}
\value{
Tidy `data.frame` containing posterior information, draws,
  and summary statistics
}
\description{
Convert list of meta-analysis results to a tidy `data.frame` of the
posterior distribution.
}
