% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-n-sensitivity-plot.R
\name{top_n_sensitivity_plot}
\alias{top_n_sensitivity_plot}
\title{Plot the top N sensitivity values}
\usage{
top_n_sensitivity_plot(
  sensitivity_plot_data,
  y_var,
  metric,
  n = 10,
  scales = "free_y"
)
}
\arguments{
\item{sensitivity_plot_data}{Formatted sensitivity results}

\item{y_var}{Response variable for which to plot sensitivity (character)}

\item{metric}{Sensitivity metric to plot, as _symbol_ (e.g. `elasticiity`,
`pvar`)}

\item{n}{Show this many values per panel}
}
\value{
`ggplot` object of plot
}
\description{
Plot the top N sensitivity values
}
\author{
Alexey Shiklomanov
}
