% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_numeric}
\alias{try_numeric}
\title{Try to convert a value to numeric. If unsuccessful, return the
original value.}
\usage{
try_numeric(x)
}
\arguments{
\item{x}{Vector to be converted}
}
\value{
`as.numeric(x)` if successful, or unaltered `x` if not
}
\description{
Try to convert a value to numeric. If unsuccessful, return the
original value.
}
