% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodel-rtm-2s.R
\name{ts_mubar_star}
\alias{ts_mubar_star}
\title{Average inverse optical depth for diffuse radiation: Two-stream}
\usage{
ts_mubar_star(orient_factor, lai, cai = 1)
}
\arguments{
\item{orient_factor}{Orient factor (npft)}

\item{lai}{Leaf area index of each cohort (ncoh)}

\item{cai}{Crown area of each cohort (ncoh)}
}
\description{
With correction for finite crown area.
}
\author{
Alexey Shiklomanov
}
