% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodel-rtm-2s.R
\name{ts_tau_diffuse}
\alias{ts_tau_diffuse}
\title{Transmissivity of diffuse radiation: Two-stream}
\usage{
ts_tau_diffuse(lai, orient_factor, lr, lt, cai)
}
\arguments{
\item{lai}{Leaf area index of each cohort (ncoh)}

\item{orient_factor}{Orient factor (npft)}

\item{cai}{Crown area of each cohort (ncoh)}
}
\description{
Transmissivity of diffuse radiation: Two-stream
}
\author{
Alexey Shiklomanov
}
