% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodel-rtm-2s.R
\name{two_stream}
\alias{two_stream}
\title{R implementation of ED2 two-stream radiation model}
\usage{
two_stream(ipft, L, lr, lt, orient, theta = 15, S0_beam = 0.8,
  S0_diff = 1 - S0_beam, alb_ground = 0.1)
}
\arguments{
\item{czen}{Cosine of angle of incidence (`cosaio`)}

\item{iota_g}{Ground (soil + snow) albedo}

\item{pft}{PFT identities of each cohort, as integer (ncoh)}

\item{lai}{Leaf area index of each cohort (ncoh)}

\item{wai}{Wood area index of each cohort (ncoh)}

\item{cai}{Crown area of each cohort (ncoh)}

\item{orient_factor}{Orient factor (npft)}

\item{clumping_factor}{Clumping factor (npft)}

\item{leaf_reflect}{Leaf reflectance spectra (nwl * npft)}

\item{leaf_trans}{Leaf transmittance spectra (nwl * npft)}

\item{wood_reflect}{Wood reflectance spectra (nwl * npft)}

\item{wood_trans}{Wood transmittance spectra (nwl * npft)}

\item{down_sky}{Normalized diffuse solar spectrum (nwl)}

\item{down0_sky}{Normalized direct solar spectrum (nwl)}

\item{wavelengths}{Numeric vector of wavelengths to use, in nm
(nwl). Default is 400:2500.}
}
\value{
List of outputs
}
\description{
R implementation of ED2 two-stream radiation model
}
\author{
Alexey Shiklomanov
}
