% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umbs-soil.R
\name{umbs_soil}
\alias{umbs_soil}
\title{UMBS soil date for inputs to ED2}
\usage{
umbs_soil()
}
\value{
`data.frame` containing soil inputs
}
\description{
UMBS soil date for inputs to ED2
}
\author{
Alexey Shiklomanov
}
