% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_prog}
\alias{with_prog}
\title{Add progress bar to function `.f`}
\usage{
with_prog(.f, ..., .pb = NULL)
}
\arguments{
\item{.f}{Function call to which to add a progress bar.}

\item{...}{Arguments to `.f`}

\item{.pb}{Progress bar object}
}
\value{
Function call that includes a call to `pb$tick()`.
}
\description{
Add progress bar to function `.f`
}
