% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_run_matrix.R
\name{workflow_matrix}
\alias{workflow_matrix}
\title{Create a `data.frame` of workflows}
\usage{
workflow_matrix(path = fs::path("analysis", "data", "model_output",
  "workflows"), since = "1970-01-01", upto = "2050-01-01")
}
\arguments{
\item{path}{Path to local workflows. Default is
`analysis/data/model_output/workflows`.}

\item{since}{Date of earliest workflow to use. Default =
`1970-01-01`.}

\item{upto}{Date of latest workflow to use. Default = `2050-01-01`.}
}
\value{
`data.frame` containing workflow IDs and paths
}
\description{
Create a `data.frame` of workflows
}
