% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_run_matrix.R
\name{workflow_run_matrix}
\alias{workflow_run_matrix}
\title{Create a `data.frame` of workflow and run IDs and directories.
Useful as an object over which to iterate.}
\usage{
workflow_run_matrix(type = c("out", "run"), ...)
}
\arguments{
\item{type}{Type of output for paths. Either `out` for output or
`run` for run configuration.}

\item{...}{Arguments passed on to \code{workflow_matrix}
\describe{
  \item{path}{Path to local workflows. Default is
`analysis/data/model_output/workflows`.}
  \item{since}{Date of earliest workflow to use. Default =
`1970-01-01`.}
  \item{upto}{Date of latest workflow to use. Default = `2050-01-01`.}
}}
}
\value{
`data.frame` of workflow and run IDs and directory paths
}
\description{
Create a `data.frame` of workflow and run IDs and directories.
Useful as an object over which to iterate.
}
\author{
Alexey Shiklomanov
}
