% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_info.R
\name{workflow_structures}
\alias{workflow_structures}
\title{Retrieve table of model structure information for each workflow}
\usage{
workflow_structures(...)
}
\arguments{
\item{...}{Arguments to [workflow_matrix()].}
}
\value{
`data.frame` of workflow ID and information about model structure
}
\description{
Retrieve table of model structure information for each workflow
}
\author{
Alexey Shiklomanov
}
