% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodel-rtm.R
\name{tau_direct}
\alias{tau_direct}
\title{Transmissivity to direct radiation}
\usage{
tau_direct(L, theta, orient, C = 1)
}
\arguments{
\item{L}{Leaf (or total) area index}

\item{theta}{Projection angle (degrees)}

\item{orient}{Leaf orientation factor (-1, 1)}

\item{C}{Crown area index (default = 1)}
}
\description{
Transmissivity to direct radiation
}
\author{
Alexey Shiklomanov
}
